move some tests to excal app and fix some
